refactor: remove insufficient funds warning; break down main component
